Move CatalogConfigEndpointContributor to `runtime/service`
Following up on #4926:

* `CatalogConfigEndpointContributor` has no use cases in `polaris-core`.

* `CatalogConfigEndpointContributor` is effectively an SPI for the IRC implementation service code (it is not a "core" SPI)

* No changes to existing module dependencies.
