+++ runtime/service/src/main/java/org/apache/polaris/service/catalog/config/CatalogConfigHandler.java
@@ -40,9 +40,9 @@ import org.apache.polaris.core.persistence.ResolvedPolarisEntity;
 import org.apache.polaris.core.persistence.resolver.Resolver;
 import org.apache.polaris.core.persistence.resolver.ResolverFactory;
 import org.apache.polaris.core.persistence.resolver.ResolverStatus;
-import org.apache.polaris.core.rest.CatalogConfigEndpointContributor;
 import org.apache.polaris.core.rest.NamespaceUtils;
 import org.apache.polaris.service.catalog.CatalogPrefixParser;
+import org.apache.polaris.service.catalog.spi.CatalogConfigEndpointContributor;
 
 @RequestScoped
 public class CatalogConfigHandler {
