+++ runtime/service/src/main/java/org/apache/polaris/service/catalog/spi/CatalogConfigEndpointContributor.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.polaris.core.rest;
+package org.apache.polaris.service.catalog.spi;
 
 import java.util.Set;
 import org.apache.iceberg.rest.Endpoint;
